removed styleguidekits and added specific files we use
