+++ patternlab-config.json
@@ -6,8 +6,8 @@
       "data" : "./source/_data/",
       "meta": "./source/_meta/",
       "annotations" : "./source/_annotations/",
-      "styleguide" : "./styleguidekit-assets-altinn/dist/",
-      "patternlabFiles" : "./styleguidekit-mustache-altinn/views/",
+      "styleguide" : "./source/styleguidekit/",
+      "patternlabFiles" : "./source/styleguidekit/mustache/views/",
       "js" : "./source/js/",
       "images" : "./source/images/",
       "fonts" : "./source/fonts/",
@@ -32,7 +32,9 @@
     "maler-portal",
     "sider-portal",
     "maler-altinnett",
-    "sider-altinnett"
+    "sider-altinnett",
+    "maler-brreg",
+    "sider-brreg"
   ],
   "defaultPattern": "all",
   "defaultShowPatternInfo": false,
