+++ source/styleguidekit/mustache/views/partials/patternSection.mustache
