+++ source/styleguidekit/mustache/views/partials/patternSectionSubtype.mustache
