+++ source/styleguidekit/styleguide/css/styleguide.css
