--- styleguidekit-assets-altinn/.bowerrc
@@ -1,3 +1,0 @@
-{
-	"directory": "src/bower_components"
-}
