--- styleguidekit-assets-altinn/.editorconfig
@@ -1,8 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-indent_style = space
-trim_trailing_whitespace = false
-end_of_line = lf
-insert_final_newline = true
