--- styleguidekit-assets-altinn/.gitignore
@@ -1,4 +1,0 @@
-node_modules/*
-src/bower_components/*
-.sass-cache/*
-.DS_Store
