--- styleguidekit-assets-altinn/.jshintrc
@@ -1,4 +1,0 @@
-{
-    "loopfunc": true,
-    "sub": true
-}
