--- styleguidekit-assets-altinn/bower.json
@@ -1,35 +1,0 @@
-{
-  "name": "styleguidekit-assets-default",
-  "version": "0.1.0",
-  "homepage": "https://github.com/pattern-lab/styleguidekit-assets-default",
-  "authors": [
-    "Dave Olsen <dmolsen@gmail.com>"
-  ],
-  "license": "MIT",
-  "private": true,
-  "dependencies": {
-    "classlist": "~2014.7.23",
-    "jquery": "~2.1.1",
-    "jwerty": "~0.3.2",
-    "prism": "gh-pages",
-    "typeahead.js": "~0.10.4",
-    "hogan": "~3.0.2",
-    "script.js": "~2.5.7",
-    "eventEmitter": "~4.2.7"
-  },
-  "overrides": {
-    "hogan": {
-      "main": "web/builds/3.0.2/hogan-3.0.2.js"
-    },
-    "script.js": {
-      "main": "dist/script.js"
-    },
-    "jwerty": {
-      "main": "jwerty.js"
-    },
-    "prism": {
-      "main": "prism.js"
-    }
-  },
-  "devDependencies": {}
-}
