--- styleguidekit-assets-altinn/composer.json
@@ -1,40 +1,0 @@
-{
-	"name":             "pattern-lab/styleguidekit-assets-default",
-	"description":      "The assets for the default StyleguideKit for Pattern Lab. Contains styles and mark-up for Pattern Lab's front-end.",
-	"keywords":         ["mustache", "pattern lab", "styleguide"],
-	"homepage":         "http://patternlab.io",
-	"license":          "MIT",
-	"type":             "patternlab-styleguidekit",
-	"authors": [
-		{
-			"name":         "Dave Olsen",
-			"email":        "dmolsen@gmail.com",
-			"homepage":     "http://dmolsen.com",
-			"role":         "Lead Developer"
-		},
-		{
-			"name":         "Brad Frost",
-			"homepage":     "http://bradfrostweb.com",
-			"role":         "Creator"
-		}
-	],
-	"support": {
-		"issues":         "https://github.com/pattern-lab/styleguidekit-assets-default/issues",
-		"wiki":           "http://patternlab.io/docs/",
-		"source":         "https://github.com/pattern-lab/styleguidekit-assets-default/releases"
-	},
-	"extra": {
-		"patternlab": {
-			"dist": {
-				"publicDir": [
-					{ "*": "*" }
-				]
-			},
-			"config": {
-				"ishMinimum": "240",
-				"ishMaximum": "2600",
-				"ishControlsHide": ["hay"]
-			}
-		}
-	}
-}
