--- styleguidekit-assets-altinn/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "styleguidekit-assets-default",
-  "version": "3.5.0",
-  "description": "The default set of front-end assets for Pattern Lab.",
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/pattern-lab/styleguidekit-assets-default.git"
-  },
-  "author": "Dave Olsen",
-  "contributors": [
-    {
-      "name": "Brian Muenzenmeyer"
-    }
-  ],
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/pattern-lab/styleguidekit-assets-default/issues"
-  },
-  "homepage": "https://github.com/pattern-lab/styleguidekit-assets-default",
-  "devDependencies": {
-    "del": "^1.1.1",
-    "gulp": "^3.8.6",
-    "gulp-autoprefixer": "^2.1",
-    "gulp-concat": "^2.3.4",
-    "gulp-file-include": "^0.13.7",
-    "gulp-imagemin": "^0.6.1",
-    "gulp-jshint": "^1.8.0",
-    "gulp-load-plugins": "^0.5.3",
-    "gulp-minify-css": "^0.3.7",
-    "gulp-rename": "^1.2.0",
-    "gulp-resolve-dependencies": "^1.0.1",
-    "gulp-ruby-sass": "^1.0.0-alpha",
-    "gulp-uglify": "^1.1.0",
-    "gulp-util": "^3.0.0",
-    "imagemin-pngcrush": "^0.1.0",
-    "main-bower-files": "^1.0.2",
-    "yargs": "^1.2.6",
-    "jquery": "3.2.1"
-  }
-}
