--- styleguidekit-assets-altinn/src/html/index.html
@@ -1,131 +1,0 @@
-<!DOCTYPE html>
-<html id="patternlab-html">
-  <head id="patternlab-head">
-
-    <title id="title">Altinn Designsystem</title>
-    <meta charset="UTF-8">
-
-    <!-- never cache patterns -->
-    <meta http-equiv="cache-control" content="max-age=0" />
-    <meta http-equiv="cache-control" content="no-cache" />
-    <meta http-equiv="expires" content="0" />
-    <meta http-equiv="expires" content="Tue, 01 Jan 1980 1:00:00 GMT" />
-    <meta http-equiv="pragma" content="no-cache" />
-
-    <!-- handle the viewport -->
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-    <!-- load the pattern lab css -->
-    <link rel="stylesheet" href="styleguide/css/styleguide.min.css" media="all" />
-    <link rel="stylesheet" href="styleguide/css/prism-typeahead.min.css" media="all" />
-
-  </head>
-  <body id="patternlab-body">
-  <!-- Pre-loader -->
-  <div class="a-sg-content-preloader">
-      <div class="a-sg-content-preloader-status">&nbsp;</div>
-  </div>
-
-    <!--Style Guide Header-->
-    <header class="sg-header" role="banner">
-        <div class="a-sg-header-top">
-            
-     <div class="a-sg-switch-dropdown">
-            <span class="a-sg-sellabel">Select project space</span>
-            <input type="hidden" name="cd-dropdown">
-            <ul class="a-sg-switch-dropdown-list">
-                <li data-value="altinn">
-                    <span id="project-altinn" class="a-sg-project-name">Altinn Designsystem</span>
-                </li>
-                <li data-value="brreg">
-                    <span id="project-brreg" class="a-sg-project-name">BRREG Designsystem</span>
-                </li>
-                <li data-value="altinnett">
-                    <span id="project-altinnett" class="a-sg-project-name">altinNETT Designsystem</span>
-                </li>
-            </ul>
-        </div>
-        <div class="a-sg-header-links">
-          <a href="https://altinn.github.io/designsystem-styleguide/">Designsystemet (Storefront)</a>
-          <a href="https://github.com/Altinn/DesignSystem">Åpne i Github </a>
-        </div>
-      </div>
-      <a href="#sg-nav-container" class="sg-nav-toggle">Meny</a>
-      <div class="sg-nav-container" id="sg-nav-container">
-        <ol class="sg-nav" id="pl-pattern-nav-target"><!-- pattern lab nav will be inserted here --></ol>
-        <div class="sg-controls" id="sg-controls"><!-- ish Controls will be inserted here --></div>
-      </div>
-    </header>
-    <!--End Style Guide Header-->
-
-    <!-- iFrame -->
-    <div id="sg-vp-wrap">
-      <div id="sg-cover"></div>
-      <div id="sg-gen-container">
-        <iframe id="sg-viewport" name="sg-viewport" sandbox="allow-same-origin allow-scripts allow-popups allow-forms"></iframe>
-        <div id="sg-rightpull-container">
-          <div id="sg-rightpull"></div>
-        </div>
-      </div>
-    </div>
-    <!--end iFrame-->
-
-    <!-- modal window -->
-    <div id="sg-modal-container" class="sg-modal anim-ready" style="bottom: -2000px">
-
-      <!-- modal close btn -->
-      <button id="sg-modal-close-btn" class="sg-modal-close-btn" title="Hide pattern info">
-          <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" width="12" height="12" viewBox="0 0 12 12">
-            <title>Close</title>
-            <path fill="#808080" d="M11.8905,9.6405 L11.8905,9.6405 L8.25,6 L11.8905,2.3595 L11.8905,2.3595 C11.9295,2.3205 11.958,2.27475 11.976,2.226 C12.0255,2.0925 11.997,1.9365 11.8905,1.82925 L10.17075,0.1095 C10.0635,0.00225 9.9075,-0.02625 9.774,0.024 C9.72525,0.042 9.6795,0.0705 9.6405,0.1095 L9.6405,0.1095 L6,3.75 L2.3595,0.1095 L2.3595,0.1095 C2.3205,0.0705 2.27475,0.042 2.226,0.024 C2.0925,-0.0255 1.9365,0.00225 1.82925,0.1095 L0.1095,1.82925 C0.00225,1.9365 -0.02625,2.0925 0.024,2.226 C0.042,2.27475 0.0705,2.3205 0.1095,2.3595 L0.1095,2.3595 L3.75,6 L0.1095,9.6405 L0.1095,9.6405 C0.0705,9.6795 0.042,9.72525 0.024,9.774 C-0.0255,9.9075 0.00225,10.0635 0.1095,10.17075 L1.82925,11.8905 C1.9365,11.99775 2.0925,12.02625 2.226,11.976 C2.27475,11.958 2.3205,11.9295 2.3595,11.8905 L2.3595,11.8905 L6,8.25 L9.6405,11.8905 L9.6405,11.8905 C9.6795,11.9295 9.72525,11.958 9.774,11.976 C9.9075,12.0255 10.0635,11.99775 10.17075,11.8905 L11.8905,10.17075 C11.99775,10.0635 12.02625,9.9075 11.976,9.774 C11.958,9.72525 11.9295,9.6795 11.8905,9.6405 L11.8905,9.6405 Z"></path>
-          </svg>
-      </button><!--end .sg-modal-close-btn-->
-
-      <div id="sg-modal-content" class="sg-modal-content"></div><!--end .sg-modal-content-->
-
-    </div><!--end .sg-modal-->
-    <!-- end modal window -->
-
-    <!-- mustache templates -->
-    <script type="text/mustache" id="pl-pattern-nav-template">
-      @@include('./partials/pattern-nav.html')
-    </script>
-
-    <script type="text/mustache" id="pl-ish-controls-template">
-      @@include('./partials/ish-controls.html')
-    </script>
-
-    <!-- the template for the modal slider -->
-    <script type="text/mustache" id="pl-panel-template-base">
-      @@include('./partials/base-template.html')
-    </script>
-
-    <!-- the tempalte for code-related tabs in the code view slider -->
-    <script type="text/mustache" id="pl-panel-template-code">
-      @@include('./partials/panel-code-template.html')
-    </script>
-
-    <!-- load the data generated by pattern lab -->
-    <script src="styleguide/data/patternlab-data.js"></script>
-    <script src="annotations/annotations.js"></script>
-
-    <!-- load the external library js -->
-    <script src="styleguide/bower_components/jquery.min.js"></script>
-    <script src="styleguide/bower_components/hogan-3.0.2.min.js"></script>
-    <script src="styleguide/bower_components/prism.min.js"></script>
-    <script src="styleguide/bower_components/jwerty.min.js"></script>
-    <script src="styleguide/bower_components/classList.min.js"></script>
-    <script src="styleguide/bower_components/typeahead.bundle.min.js"></script>
-    <script src="styleguide/bower_components/EventEmitter.min.js"></script>
-    <script src="styleguide/bower_components/script.min.js"></script>
-
-    <!-- set-up the dispatcher -->
-    <script>
-      var Dispatcher = new EventEmitter();
-    </script>
-
-    <!-- load the pattern lab viewer js -->
-    <script src="styleguide/js/patternlab-viewer.min.js"></script>
-
-  </body>
-</html>
