--- styleguidekit-assets-altinn/src/html/partials/base-template.html
@@ -1,97 +1,0 @@
-<div class="sg-pattern-extra-inner">
-
-  <!-- description, etc. -->
-  {{# descBlockExists }}
-
-    <div class="sg-pattern-extra-info">
-
-      {{# isPatternView }}
-        <!-- if this is for a single pattern provide some extra info -->
-        <div class="sg-pattern-head">
-
-          <div class="sg-pattern-breadcrumb">
-            {{# patternBreadcrumb }}
-              {{ patternType }}
-              {{# patternSubtype }}
-                / {{ patternSubtype }}
-              {{/ patternSubtype }}
-            {{/ patternBreadcrumb }}
-          </div>
-
-          <h2 class="sg-pattern-title">
-            {{ patternNameCaps }}
-            {{# patternState }}
-              <span class="sg-pattern-state {{ patternState }}">{{ patternState }}</span>
-            {{/ patternState }}
-          </h2>
-
-        </div><!--end .sg-pattern-head-->
-      {{/ isPatternView }}
-
-      {{# patternDescExists }}
-        <div class="sg-pattern-desc">
-          {{{ patternDesc }}}
-          {{# patternDescAdditions }}
-            {{{ patternDescAdditions }}}
-          {{/ patternDescAdditions }}
-        </div><!--end .sg-pattern-desc-->
-      {{/ patternDescExists }}
-
-      {{# lineageExists }}
-        <p class="sg-pattern-lineage">
-          <strong>{{ patternName }}</strong> inneholder følgende komponenter:
-          <span id="sg-code-lineage-fill">
-          {{# lineage }}
-            <a href='{{ lineagePath }}' class='{{# lineageState }}sg-pattern-state {{ lineageState }}{{/ lineageState }}' data-patternPartial='{{ lineagePattern }}'>{{ lineagePattern }}</a>{{# hasComma }}, {{/ hasComma }}
-          {{/ lineage }}
-          </span>
-        </p><!--end .sg-pattern-lineage-->
-      {{/ lineageExists }}
-
-      {{# lineageRExists }}
-        <p class="sg-pattern-lineage">
-          <strong>{{ patternName }}</strong> er inkludert følgende steder:
-          <span id="sg-code-lineager-fill">
-          {{# lineageR }}
-            <a href='{{ lineagePath }}' class='{{# lineageState }}sg-pattern-state {{ lineageState }}{{/ lineageState }}' data-patternPartial='{{ lineagePattern }}'>{{ lineagePattern }}</a>{{# hasComma }}, {{/ hasComma }}
-          {{/ lineageR }}
-          </span>
-        </p><!--end .sg-pattern-lineage-->
-      {{/ lineageRExists }}
-
-      {{# annotationExists }}
-        <div class="sg-annotations" id="sg-annotations">
-          <h2 class="sg-annotations-title">Annotations</h2>
-          <ol class="sg-annotations-list">
-            {{# annotations }}
-              <li><h4 class="sg-annotations-list-title">{{ title }}</h4> {{{ comment }}}</li>
-            {{/ annotations }}
-          </ol>
-        </div>
-      {{/ annotationExists }}
-
-    </div><!--end .sg-pattern-extra-info-->
-  {{/ descBlockExists }}
-
-  <div class="sg-pattern-extra-code">
-
-    <div id="sg-{{ patternPartial }}-tabs" class="sg-tabs">
-      <ul class="sg-tabs-list">
-        {{# panels }}
-          <li id="sg-{{ patternPartial }}-{{ id }}-tab" data-patternpartial="{{ patternPartial }}" data-panelid="{{ id }}"><a href="#sg-{{ patternPartial }}-{{ id }}-panel" class="active">{{ name }}</a></li>
-        {{/ panels }}
-      </ul><!--end .sg-tabs-list-->
-
-      <div id="sg-{{ patternPartial }}-panels" class="sg-tabs-content">
-        {{# panels }}
-          <div id="sg-{{ patternPartial }}-{{ id }}-panel"  class="sg-tabs-panel" style="display: none">
-            {{{ content }}}
-          </div>
-        {{/ panels }}
-      </div><!--end .sg-tabs-content-->
-
-    </div><!--end .sg-tabs-->
-
-  </div><!--end .sg-pattern-extra-code-->
-
-</div><!--end .sg-pattern-extra-inner-->
