--- styleguidekit-assets-altinn/src/html/partials/panel-code-template.html
@@ -1,3 +1,0 @@
-<pre class="language-markup">
-  <code id="sg-code-fill-{{ language }}" class="language-{{ language }}">{{{ code }}}</code>
-</pre>
