--- styleguidekit-assets-altinn/src/html/partials/pattern-nav.html
@@ -1,17 +1,0 @@
-{{^ ishControlsHide.views-all }}
-  <li><a href="styleguide/html/styleguide.html" class="sg-pop sg-nav-menus" style="display: none;" data-patternpartial="all">Alle komponenter</a></li>
-{{/ ishControlsHide.views-all }}
-{{# patternTypes }}
-  <li><a class="sg-acc-handle sg-nav-menus" style="display: none;">{{ patternTypeUC }}</a><ol class="sg-acc-panel">
-  {{# patternTypeItems }}
-    <li><a class="sg-acc-handle">{{ patternSubtypeUC }}</a><ol class="sg-acc-panel sg-sub-nav">
-    {{# patternSubtypeItems }}
-      <li><a href="patterns/{{ patternPath }}" class="sg-pop  {{# patternState }}sg-pattern-state {{ patternState }}{{/ patternState }}" data-patternpartial="{{ patternPartial }}">{{ patternName }}</a></li>
-    {{/ patternSubtypeItems }}
-    </ol></li>
-  {{/ patternTypeItems }}
-  {{# patternItems }}
-    <li><a href="patterns/{{ patternPath }}" class="sg-pop {{# patternState }}sg-pattern-state {{ patternState }}{{/ patternState }}" data-patternpartial="{{ patternPartial }}">{{ patternName }}</a></li>
-  {{/ patternItems }}
-  </ol></li>
-{{/ patternTypes }}
