--- styleguidekit-mustache-altinn/README.md
@@ -1,14 +1,0 @@
-![license](https://img.shields.io/github/license/pattern-lab/styleguidekit-mustache-default.svg?maxAge=2592000)
-[![Packagist](https://img.shields.io/packagist/v/pattern-lab/styleguidekit-mustache-default.svg?maxAge=2592000)](https://packagist.org/packages/pattern-lab/styleguidekit-mustache-default) [![Gitter](https://img.shields.io/gitter/room/pattern-lab/php.svg?maxAge=2592000)](https://gitter.im/pattern-lab/frontend-viewer)
-
-# Default Mustache StyleguideKit for Pattern Lab
-
-The default Mustache StyleguideKit provides the basic front-end assets when using the [Mustache PatternEngine](https://github.com/pattern-lab/patternengine-php-mustache). This StyleguideKit includes Mustache templates for the "View All" views and works with the [default StyleguideKit static assets](https://github.com/pattern-lab/styleguidekit-assets-default).
-
-## Installation
-
-The default Mustache StyleguideKit can be installed via Composer:
-
-    composer require pattern-lab/styleguidekit-mustache-default
-
-Learn more about this StyleguideKit's dependencies on [Packagist](https://packagist.org/packages/pattern-lab/styleguidekit-mustache-default).
