--- styleguidekit-mustache-altinn/composer.json
@@ -1,32 +1,0 @@
-{
-  "name":             "pattern-lab/styleguidekit-mustache-default",
-  "description":      "The default Mustache-based StyleguideKit for Pattern Lab. Contains styles and mark-up for Pattern Lab's front-end.",
-  "keywords":         ["mustache", "pattern lab", "styleguide"],
-  "homepage":         "http://patternlab.io",
-  "license":          "MIT",
-  "type":             "patternlab-styleguidekit",
-  "authors": [
-    {
-      "name":         "Dave Olsen",
-      "email":        "dmolsen@gmail.com",
-      "homepage":     "http://dmolsen.com",
-      "role":         "Lead Developer"
-    },
-    {
-      "name":         "Brad Frost",
-      "homepage":     "http://bradfrostweb.com",
-      "role":         "Creator"
-    }
-  ],
-  "support": {
-    "issues":         "https://github.com/pattern-lab/styleguidekit-mustache-default/issues",
-    "wiki":           "http://patternlab.io/docs/",
-    "source":         "https://github.com/pattern-lab/styleguidekit-mustache-default/releases"
-  },
-  "minimum-stability": "dev",
-  "require": {
-    "pattern-lab/core": "^2.0.0",
-    "pattern-lab/patternengine-mustache": "^2.0.0",
-    "pattern-lab/styleguidekit-assets-default": "^3.0.0"
-  }
-}
