--- styleguidekit-mustache-altinn/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "styleguidekit-mustache-default",
-  "description": "The default Mustache-based StyleguideKit for Pattern Lab. Contains styles and mark-up for Pattern Lab's front-end.",
-  "version": "3.1.0",
-  "dependencies": {},
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/pattern-lab/styleguidekit-mustache-default.git"
-  },
-  "bugs": "https://github.com/pattern-lab/styleguidekit-mustache-default/issues",
-  "authors": [
-    {
-      "name": "Dave Olsen",
-      "email": "dmolsen@gmail.com",
-      "homepage": "http://dmolsen.com",
-      "role": "Lead Developer"
-    },
-    {
-      "name": "Brad Frost",
-      "homepage": "http://bradfrostweb.com",
-      "role": "Creator"
-    },
-    {
-      "name": "Brian Muenzenmeyer",
-      "homepage": "http://brianmuenzenmeyer.com",
-      "role": "Developer"
-    }
-  ],
-  "license": "MIT"
-}
